Move activity lifecycle classes to the activity package
