+++ app/src/main/java/org/simple/clinic/activity/RxTheActivityLifecycle.kt
@@ -1,10 +1,7 @@
-package org.simple.clinic.widgets
+package org.simple.clinic.activity
 
 import android.app.Activity
-import android.app.Application.ActivityLifecycleCallbacks
-import android.os.Bundle
 import io.reactivex.Observable
-import org.simple.clinic.activity.TheActivity
 
 class RxTheActivityLifecycle internal constructor(private val events: Observable<TheActivityLifecycle>) {
 
@@ -56,20 +53,4 @@ class RxTheActivityLifecycle internal constructor(private val events: Observable
       return RxTheActivityLifecycle(lifecycleEvents)
     }
   }
-}
-
-abstract class SimpleActivityLifecycleCallbacks : ActivityLifecycleCallbacks {
-  override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {}
-
-  override fun onActivityStarted(activity: Activity) {}
-
-  override fun onActivityResumed(activity: Activity) {}
-
-  override fun onActivityPaused(activity: Activity) {}
-
-  override fun onActivityStopped(activity: Activity) {}
-
-  override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
-
-  override fun onActivityDestroyed(activity: Activity) {}
 }
