+++ app/src/main/java/org/simple/clinic/activity/SimpleActivityLifecycleCallbacks.kt
@@ -1,0 +1,21 @@
+package org.simple.clinic.activity
+
+import android.app.Activity
+import android.app.Application
+import android.os.Bundle
+
+abstract class SimpleActivityLifecycleCallbacks : Application.ActivityLifecycleCallbacks {
+  override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {}
+
+  override fun onActivityStarted(activity: Activity) {}
+
+  override fun onActivityResumed(activity: Activity) {}
+
+  override fun onActivityPaused(activity: Activity) {}
+
+  override fun onActivityStopped(activity: Activity) {}
+
+  override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
+
+  override fun onActivityDestroyed(activity: Activity) {}
+}
