+++ app/src/main/java/org/simple/clinic/activity/TheActivityLifecycle.kt
@@ -1,4 +1,6 @@
-package org.simple.clinic.widgets
+package org.simple.clinic.activity
+
+import org.simple.clinic.widgets.UiEvent
 
 sealed class TheActivityLifecycle : UiEvent {
