+++ app/src/test/java/org/simple/clinic/activity/TheActivityControllerTest.kt
@@ -39,8 +39,6 @@ import org.simple.clinic.util.Just
 import org.simple.clinic.util.None
 import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RxErrorsRule
-import org.simple.clinic.util.toOptional
-import org.simple.clinic.widgets.TheActivityLifecycle
 import org.simple.clinic.widgets.UiEvent
 import org.threeten.bp.Instant
 import java.util.concurrent.TimeUnit
